bricks_production/requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
Flask-WTF==1.1.1
Flask-Login==0.6.2
Werkzeug==2.3.7
email-validator==2.0.0
python-dotenv==1.0.0
stripe==6.1.0
Pillow==10.0.1
flask-bcrypt==1.0.1 